+++ preview/FormFillerPreview.tsx
@@ -240,6 +240,15 @@ const FormFillerPreview = (): JSX.Element => {
     );
   };
 
+  const handleFocus = (refElement: HTMLElement, isStepView: boolean, stepIndex: number | undefined): void => {
+    const shouldFocus = isStepView && stepIndex && stepIndex > 0;
+    if (shouldFocus) {
+      const focusableElements = 'input, select, textarea, button, a[href], [tabindex]:not([tabindex="-1"])';
+      const firstFocusable = refElement.querySelector<HTMLElement>(focusableElements);
+      firstFocusable?.focus();
+    }
+  };
+
   if (!questionnaire || !questionnaireResponse) return <div>{'loading...'}</div>;
   return (
     <Provider store={store}>
@@ -286,9 +295,9 @@ const FormFillerPreview = (): JSX.Element => {
                     return <HelpButton opening={opening} />;
                   }}
                   onFieldsNotCorrectlyFilledOut={() => {}}
-                  onFormViewChange={() => {}}
-
-                  // onStepChange={(newIndex: number): void => setStepIndex(newIndex)}
+                  onFormViewChange={(containerRef, stepIndex) => {
+                    handleFocus(containerRef, false, stepIndex);
+                  }}
                 />
               </div>
             ) : (